pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blue-noise-stippling"
version = "0.1.0"
description = "Blue-noise stippling using modified void-and-cluster algorithm"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "Pillow>=8.0.0",
    "matplotlib>=3.3.0",
    "imageio>=2.9.0",
]

[project.scripts]
stipple = "stipple.runner:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = ["C0111", "R0903"]

